+++ template/node/src/service.rs
@@ -20,6 +20,7 @@ use sp_timestamp::InherentError;
 use sc_telemetry::{Telemetry, TelemetryWorker};
 use sc_cli::SubstrateCli;
 use futures::StreamExt;
+use sp_core::U256;
 
 use crate::cli::Cli;
 #[cfg(feature = "manual-seal")]
@@ -30,6 +31,7 @@ native_executor_instance!(
 	pub Executor,
 	frontier_template_runtime::api::dispatch,
 	frontier_template_runtime::native_version,
+	frame_benchmarking::benchmarking::HostFunctions,
 );
 
 type FullClient = sc_service::TFullClient<Block, RuntimeApi, Executor>;
@@ -82,18 +84,20 @@ impl ProvideInherentData for MockTimestampInherentDataProvider {
 	}
 }
 
-pub fn open_frontier_backend(config: &Configuration) -> Result<Arc<fc_db::Backend<Block>>, String> {
+pub fn frontier_database_dir(config: &Configuration) -> std::path::PathBuf {
 	let config_dir = config.base_path.as_ref()
 		.map(|base_path| base_path.config_dir(config.chain_spec.id()))
 		.unwrap_or_else(|| {
 			BasePath::from_project("", "", &crate::cli::Cli::executable_name())
 				.config_dir(config.chain_spec.id())
 		});
-	let database_dir = config_dir.join("frontier").join("db");
+	config_dir.join("frontier").join("db")
+}
 
+pub fn open_frontier_backend(config: &Configuration) -> Result<Arc<fc_db::Backend<Block>>, String> {
 	Ok(Arc::new(fc_db::Backend::<Block>::new(&fc_db::DatabaseSettings {
 		source: fc_db::DatabaseSettingsSrc::RocksDb {
-			path: database_dir,
+			path: frontier_database_dir(&config),
 			cache_size: 0,
 		}
 	})?))
@@ -155,6 +159,10 @@ pub fn new_partial(config: &Configuration, #[allow(unused_variables)] cli: &Cli)
 			.register_provider(MockTimestampInherentDataProvider)
 			.map_err(Into::into)
 			.map_err(sp_consensus::error::Error::InherentData)?;
+		inherent_data_providers
+			.register_provider(pallet_dynamic_fee::InherentDataProvider(U256::from(cli.run.target_gas_price)))
+			.map_err(Into::into)
+			.map_err(sp_consensus::Error::InherentData)?;
 
 		let frontier_block_import = FrontierBlockImport::new(
 			client.clone(),
@@ -182,6 +190,11 @@ pub fn new_partial(config: &Configuration, #[allow(unused_variables)] cli: &Cli)
 	}
 
 	#[cfg(feature = "aura")] {
+		inherent_data_providers
+			.register_provider(pallet_dynamic_fee::InherentDataProvider(U256::from(cli.run.target_gas_price)))
+			.map_err(Into::into)
+			.map_err(sp_consensus::Error::InherentData)?;
+
 		let (grandpa_block_import, grandpa_link) = sc_finality_grandpa::block_import(
 			client.clone(),
 			&(client.clone() as Arc<_>),