+++ client/rpc/CHANGELOG.md
@@ -3,3 +3,5 @@
 ## Unreleased
 
 * `EthPubSubApi::new` takes an additional `overrides` parameter.
+* Fix `estimate_gas` inaccurate issue.
+* Use pallet-ethereum 3.0.0-dev.