+++ frame/dynamic-fee/CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog for `pallet-dynamic-fee`
 
 ## Unreleased
+
+* Uses unreleased pallet-evm 5.0.0-dev